--- h/BranchAction.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "Action.h"
-
-ActionPtr branchAction(ActionPtr leftInputIn, ActionPtr actionIn, ActionPtr rightInputIn);
