--- h/FunctionAction.h
@@ -1,7 +1,0 @@
-
-#pragma once
-
-#include "Action.h"
-
-ActionPtr functionAction(ActionPtr actionIn, Type inLeftTypeIn, Type inRightTypeIn, int dataSizeIn, string textIn);
-
