--- h/IfAction.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "Action.h"
-
-ActionPtr ifAction(ActionPtr conditionIn, ActionPtr ifActionIn);
-
-
