--- h/ListAction.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "Action.h"
-#include <list>
-
-ActionPtr listAction(std::list<ActionPtr>& actionsIn);
