--- h/LoopAction.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "Action.h"
-
-ActionPtr loopAction(ActionPtr conditionIn, ActionPtr loopActionIn);
-
-
