+++ h/MakeTupleAction.h
@@ -2,6 +2,3 @@
 
 #include "Action.h"
 #include <vector>
-
-ActionPtr makeTupleAction(std::vector<ActionPtr>& sourceActionsIn);
-
