+++ src/Actions/BranchAction.cpp
@@ -1,5 +1,5 @@
-#include "../h/BranchAction.h"
-#include "../h/ErrorHandler.h"
+#include "../../h/Action.h"
+#include "../../h/ErrorHandler.h"
 
 class BranchAction: public Action
 {
