+++ src/Actions/FunctionAction.cpp
@@ -1,7 +1,7 @@
 
-#include "../h/FunctionAction.h"
-#include "../h/ErrorHandler.h"
-#include "../h/StackFrame.h"
+#include "../../h/Action.h"
+#include "../../h/ErrorHandler.h"
+#include "../../h/StackFrame.h"
 
 #include <cstring> //for memcpy
 using std::memcpy;
