+++ src/ElementList.cpp
@@ -3,7 +3,6 @@
 #include "../h/OperatorElement.h"
 #include "../h/LiteralElement.h"
 #include "../h/ErrorHandler.h"
-#include "../h/ListAction.h"
 #include "../h/StackFrame.h"
 #include "../h/TupleElement.h"
